Update custom editor webview when document changes
The webview that is the custom editor now updates automatically
when the document changes. So, if you simultaneously open the
mce using the stock text editor and modify the mce file that
way, the custom editor now autoamtically reflects the changed
content. Also, clicking the buttin now shows the new content in the
custom editor; you con't have to close the editor and re-open it.