Test/demonstrate different Webview persistence mechanisms
You have your React state, which by default, is destroyed when the
view is hidden.

You have browser sessionStorage and localStorage, which are parts
of the 'window' Web API available to javascript running in browsers.

Session storage, unlike React state, survives quite a bit:
* view visibility state change (the view being hidden and shown)
* view being closed and reopened
* app referesh (CTRL-R in VS Code)
It does not, however, survive an app restart.

Local storage provides the persistence of session storage plus
it survives an app restart--i.e., the data is stored to disk.